backend-py/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ved-io-backend"
version = "1.0.0"
description = "Ved.io Drama Generator — Python Backend (FastAPI)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "ffmpeg-python>=0.2.0",
    "apscheduler>=3.10.4",
    "tenacity>=8.3.0",
    "cachetools>=5.3.3",
    "pillow>=10.3.0",
    "slowapi>=0.1.9",
    "python-jose[cryptography]>=3.3.0",
    "anyio>=4.4.0",
    "pyyaml>=6.0.1",
    "google-generativeai>=0.7.0",
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "black>=24.4.2",
    "ruff>=0.4.7",
    "mypy>=1.10.0",
    "bandit>=1.7.9",
    "pre-commit>=3.7.0",
    "factory-boy>=3.3.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "alembic/*", ".venv/*"]
